2024.04.22 commit
고객 리뷰 기능(리뷰 달기, 리뷰 삭제하기) 구현
ReviewDAO 생성
리뷰 등록,삭제시 주문상태 변경 기능 구현
주문 상태 변경 기능 통합(emp, customer 분리돼있던것)
평점 별모양 달기 css 구현 및 적용

+++ shop/src/main/webapp/customer/deleteCustomerAction.jsp
@@ -1,3 +1,4 @@
+<%@page import="java.net.URLEncoder"%>
 <%@page import="shop.dao.CustomerDAO"%>
 <%@page import="java.util.HashMap"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
@@ -14,11 +15,24 @@
 	// 해당 jsp 바로 실행했을 경우
 	if(customerId == null && customerPw == null) {
 		response.sendRedirect("/shop/customer/goods/customerGoodsList.jsp");
+		return;
 	}
 %>
 <%
-	// 회원 삭제
-	int deleteCustomerRow = CustomerDAO.deleteCustomer(customerId, customerPw);
+	// 회원정보 일치 검증
+	boolean canDeleteCustomer = CustomerDAO.checkCustomerIdPw(customerId, customerPw);
+
+	int deleteCustomerRow = 0;
+	// 회원 정보 일치할 경우
+	if(canDeleteCustomer) {
+		// 회원 삭제
+		deleteCustomerRow = CustomerDAO.deleteCustomer(customerId, customerPw);	
+	} else {
+		// 회원 정보 일치하지 않을 경우
+		String msg = URLEncoder.encode("회원 탈퇴에 실패헀습니다. 다시 입력 해주세요", "UTF-8");
+		response.sendRedirect("/shop/customer/deleteCustomerForm.jsp?msg=" + msg);
+		return;
+	}
 	
 	System.out.println("deleteCustomerAction - deleteCustomerRow = " + deleteCustomerRow);
 	// 회원 탈퇴시 로그인돼있던 회원 세션 초기화
